Fixed minor bugs and improved cone fitting
git-svn-id: http://code.in.tum.de/svn/knowrob/trunk@795 63041b90-49fb-4ca4-89d8-2ccc525231da

+++ knowrob_cad_parser/src/edu/tum/cs/vis/model/util/Vertex.java
@@ -83,7 +83,7 @@ public class Vertex extends Point3f {
 	@Override
 	public int hashCode() {
 		return Float.valueOf(x).hashCode() ^ Float.valueOf(y).hashCode()
-				^ Float.valueOf(z).hashCode() ^ Float.valueOf(pointarea).hashCode()^normalVector.hashCode();
+				^ Float.valueOf(z).hashCode() ^ Double.valueOf(pointarea).hashCode()^normalVector.hashCode();
 	}
 
 	/**
